+++ src/Products/OFSP/help/Response.py
@@ -110,7 +110,7 @@ class Response:
 
         '''
 
-    def setCookie(name,value,**kw):
+    def setCookie(name, value, quoted=True, **kw):
         '''
 
         Set an HTTP cookie on the browser
@@ -119,6 +119,10 @@ class Response:
         cookie-enabled browsers with a key "name" and value
         "value". This overwrites any previously set value for the
         cookie in the Response object.
+
+        By default, the cookie value will be enclosed in double quotes.
+        To suppress the double quotes you can pass the "quoted" argument
+        with a False value such as False or 0.
 
         Permission -- Always available