+++ src/ZPublisher/tests/testHTTPResponse.py
@@ -34,8 +34,9 @@ class HTTPResponseTests(unittest.TestCase):
         response = self._makeOne()
         response.setCookie('foo', 'bar')
         cookie = response.cookies.get('foo', None)
-        self.assertEqual(len(cookie), 1)
+        self.assertEqual(len(cookie), 2)
         self.assertEqual(cookie.get('value'), 'bar')
+        self.assertEqual(cookie.get('quoted'), True)
 
         cookies = response._cookie_list()
         self.assertEqual(len(cookies), 1)
@@ -49,6 +50,7 @@ class HTTPResponseTests(unittest.TestCase):
         self.failUnless(cookie)
         self.assertEqual(cookie.get('value'), 'bar')
         self.assertEqual(cookie.get('expires'), EXPIRES)
+        self.assertEqual(cookie.get('quoted'), True)
 
         cookies = response._cookie_list()
         self.assertEqual(len(cookies), 1)
@@ -59,9 +61,10 @@ class HTTPResponseTests(unittest.TestCase):
         response = self._makeOne()
         response.setCookie('foo', 'bar', domain='example.com')
         cookie = response.cookies.get('foo', None)
-        self.assertEqual(len(cookie), 2)
+        self.assertEqual(len(cookie), 3)
         self.assertEqual(cookie.get('value'), 'bar')
         self.assertEqual(cookie.get('domain'), 'example.com')
+        self.assertEqual(cookie.get('quoted'), True)
 
         cookies = response._cookie_list()
         self.assertEqual(len(cookies), 1)
@@ -72,9 +75,10 @@ class HTTPResponseTests(unittest.TestCase):
         response = self._makeOne()
         response.setCookie('foo', 'bar', path='/')
         cookie = response.cookies.get('foo', None)
-        self.assertEqual(len(cookie), 2)
+        self.assertEqual(len(cookie), 3)
         self.assertEqual(cookie.get('value'), 'bar')
         self.assertEqual(cookie.get('path'), '/')
+        self.assertEqual(cookie.get('quoted'), True)
 
         cookies = response._cookie_list()
         self.assertEqual(len(cookies), 1)
@@ -84,9 +88,10 @@ class HTTPResponseTests(unittest.TestCase):
         response = self._makeOne()
         response.setCookie('foo', 'bar', comment='COMMENT')
         cookie = response.cookies.get('foo', None)
-        self.assertEqual(len(cookie), 2)
+        self.assertEqual(len(cookie), 3)
         self.assertEqual(cookie.get('value'), 'bar')
         self.assertEqual(cookie.get('comment'), 'COMMENT')
+        self.assertEqual(cookie.get('quoted'), True)
 
         cookies = response._cookie_list()
         self.assertEqual(len(cookies), 1)
@@ -96,9 +101,10 @@ class HTTPResponseTests(unittest.TestCase):
         response = self._makeOne()
         response.setCookie('foo', 'bar', secure='SECURE')
         cookie = response.cookies.get('foo', None)
-        self.assertEqual(len(cookie), 2)
+        self.assertEqual(len(cookie), 3)
         self.assertEqual(cookie.get('value'), 'bar')
         self.assertEqual(cookie.get('secure'), 'SECURE')
+        self.assertEqual(cookie.get('quoted'), True)
 
         cookies = response._cookie_list()
         self.assertEqual(len(cookies), 1)
@@ -108,9 +114,10 @@ class HTTPResponseTests(unittest.TestCase):
         response = self._makeOne()
         response.setCookie('foo', 'bar', secure='')
         cookie = response.cookies.get('foo', None)
-        self.assertEqual(len(cookie), 2)
+        self.assertEqual(len(cookie), 3)
         self.assertEqual(cookie.get('value'), 'bar')
         self.assertEqual(cookie.get('secure'), '')
+        self.assertEqual(cookie.get('quoted'), True)
 
         cookies = response._cookie_list()
         self.assertEqual(len(cookies), 1)
@@ -129,9 +136,10 @@ class HTTPResponseTests(unittest.TestCase):
         response = self._makeOne()
         response.setCookie('foo', 'bar', http_only=True)
         cookie = response.cookies.get('foo', None)
-        self.assertEqual(len(cookie), 2)
+        self.assertEqual(len(cookie), 3)
         self.assertEqual(cookie.get('value'), 'bar')
         self.assertEqual(cookie.get('http_only'), True)
+        self.assertEqual(cookie.get('quoted'), True)
 
         cookie_list = response._cookie_list()
         self.assertEqual(len(cookie_list), 1)
@@ -141,13 +149,26 @@ class HTTPResponseTests(unittest.TestCase):
         response = self._makeOne()
         response.setCookie('foo', 'bar', http_only=False)
         cookie = response.cookies.get('foo', None)
-        self.assertEqual(len(cookie), 2)
+        self.assertEqual(len(cookie), 3)
         self.assertEqual(cookie.get('value'), 'bar')
         self.assertEqual(cookie.get('http_only'), False)
+        self.assertEqual(cookie.get('quoted'), True)
 
         cookie_list = response._cookie_list()
         self.assertEqual(len(cookie_list), 1)
         self.assertEqual(cookie_list[0], 'Set-Cookie: foo="bar"')
+
+    def test_setCookie_unquoted(self):
+        response = self._makeOne()
+        response.setCookie('foo', 'bar', quoted=False)
+        cookie = response.cookies.get('foo', None)
+        self.assertEqual(len(cookie), 2)
+        self.assertEqual(cookie.get('value'), 'bar')
+        self.assertEqual(cookie.get('quoted'), False)
+
+        cookie_list = response._cookie_list()
+        self.assertEqual(len(cookie_list), 1)
+        self.assertEqual(cookie_list[0], 'Set-Cookie: foo=bar')
 
     def test_expireCookie1160(self):
         # Verify that the cookie is expired even if an expires kw arg is passed